+++ gleam_stdlib/gen/src/float.erl
@@ -1,7 +1,7 @@
 -module(float).
 -compile(no_auto_import).
 
--export([parse/1, to_string/1, ceiling/1, floor/1]).
+-export([parse/1, to_string/1, ceiling/1, floor/1, round/1]).
 
 parse(A) ->
     gleam__stdlib:parse_float(A).
@@ -14,3 +14,6 @@ ceiling(A) ->
 
 floor(A) ->
     math:floor(A).
+
+round(A) ->
+    erlang:round(A).